moved rendermovies to utils.js
